ci(github): use google-github-actions/setup-gcloud@v0
The `master` branch in the source repository has been renamed to `main`, but since `v0` is currently pointing to the latest release, we'll use that from now on.

In addition, the authentication functionality is deprecated in the `setup-gcloud` action, so we have to add a new step using the [`auth` action](https://github.com/google-github-actions/auth).