style(commitlint): extend `scope-enum` with additional values
- start to document the accepted scopes
- add `commitlint`, `github`, and `make` scopes

Why?

My personal experience has shown me that I tend to forget my own conventions of which scopes to use over time, which is why I think the scopes should be documented somewhere.

I sometimes want to omit scopes entirely, but while that might be a good choice for a leaf repository which isn't used as an upstream template, the "terragrunt" repository _is_ a template and the changes introduced by it should be clearly marked in the history of downstream projects.