ci(github): re-trigger workflow on push to release branch
The goal of the previous change was to update the release branch itself (and re-trigger another workflow run to give a final feedback) whenever automatic fixes were made or dependency graphs were updated. Instead, we continued to try to create a new pull request.

This should fix the previous change which introduced a `fix` job on release PRs.