ci(github): run `make test`, but only in changed directories
This change reduces the number of unnecessary TFLint executions and enables automated Terraform and Terragrunt test executions.

It does this by replacing the `lint` goal with the `test` goal in the CI workflow and also setting `SKIP_UNCHANGED=true` to enter only subdirectories which contain changes compared to the main branch.