fix(terragrunt-aws): raise precedence of layer variables
This change always gives precedence to automatic input variables based on directory names over input variables defined in terragrunt.yml.

For example, a directory layer called `region` should be able to override the `region` input variables set in an outer terragrunt.yml.

Caveats: This change gives a higher precedence to _all_ layer inputs. It is not possible to set `region` in `terragrunt.yml`, override it with an automatic directory layer called `region`, and then override it again in a `terragrunt.yml` within that directory.